requirements.txt
# Django Stuff
Django==4.2.11
djangorestframework==3.14.0
